+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/config/WheelsConfigs.java
@@ -16,9 +16,9 @@ import org.firstinspires.ftc.teamcode.wheels.WheelMotor;
 import org.firstinspires.ftc.teamcode.wheels.Wheels;
 
 public class WheelsConfigs extends Configs {
-    public final static double DERATE = 0.875;
-    public final static double WC_MAX_RATE = 2.655d * DERATE;
-    public final static double WC_TICKS_PER_MM = 0.92d;
+    public final static float DERATE = 0.875f;
+    public final static float WC_MAX_RATE = 2.655f * DERATE;
+    public final static float WC_TICKS_PER_MM = 0.92f;
 
     public WheelsConfigs(HardwareMap map, Telemetry telemetry, BOT bot) {
         super(map, telemetry, bot);
@@ -49,9 +49,9 @@ public class WheelsConfigs extends Configs {
             case WestCoast:
                 motors = new WheelMotor[4];
                 motors[0] = new WheelMotor("ML1", MOTOR_SIDE.LEFT, true,
-                        new PIDParams(0.38d, 1.00d, 0.0d, WC_MAX_RATE, WC_TICKS_PER_MM));
+                        new PIDParams(0.38f, 1.00f, 0.0f, WC_MAX_RATE, WC_TICKS_PER_MM));
                 motors[1] = new WheelMotor("MR1", MOTOR_SIDE.RIGHT, false,
-                        new PIDParams(0.29d, 0.25d, 0.0d, WC_MAX_RATE, WC_TICKS_PER_MM));
+                        new PIDParams(0.29f, 0.25f, 0.0f, WC_MAX_RATE, WC_TICKS_PER_MM));
                 motors[2] = new WheelMotor("ML2", MOTOR_SIDE.LEFT, true);
                 motors[3] = new WheelMotor("MR2", MOTOR_SIDE.RIGHT, false);
                 config = new WheelsConfig(DRIVE_TYPE.TANK, motors);
@@ -59,7 +59,7 @@ public class WheelsConfigs extends Configs {
             case Mecanum:
                 motors = new WheelMotor[4];
                 motors[0] = new WheelMotor("FL", MOTOR_SIDE.LEFT, MOTOR_END.FRONT, false,
-                        new PIDParams(0.1d, 0.01d, 0.0d, 1.0d, 1.0d));
+                        new PIDParams(0.1f, 0.01f, 0.0f, 1.0f, 1.0f));
                 motors[1] = new WheelMotor("BL", MOTOR_SIDE.LEFT, MOTOR_END.BACK, false);
                 motors[2] = new WheelMotor("FR", MOTOR_SIDE.RIGHT, MOTOR_END.FRONT, true);
                 motors[3] = new WheelMotor("BR", MOTOR_SIDE.RIGHT, MOTOR_END.BACK, true);
